pyproject.toml
[project]
name = "ai-update-radar"
version = "0.1.0"
description = "Detect, evaluate, and experiment with meaningful AI updates"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.0",      # RSS/Atom パーサー
    "httpx>=0.27.0",          # HTTP クライアント
    "pyyaml>=6.0",            # YAML パーサー
    "rich>=13.0.0",           # リッチ出力
    "typer>=0.12.0",          # CLI
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[project.scripts]
radar = "collectors.cli:app"

[tool.setuptools.packages.find]
include = ["collectors*", "evaluators*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
